reformat .sol files

+++ contracts/interfaces/IMark2Market.sol
@@ -2,19 +2,16 @@
 
 pragma solidity >=0.8.0 <0.9.0;
 
-interface IMark2Market   {
+interface IMark2Market {
     struct ActivesPrices {
         address addr;
         string name;
         string symbol;
-        uint decimals;
+        uint256 decimals;
         uint256 price;
         uint256 bookValue;
         uint256 liquidationValue;
-
     }
 
-    function activesPrices () external view returns (ActivesPrices[] memory ) ;
-
-
+    function activesPrices() external view returns (ActivesPrices[] memory);
 }
